refactor: extract send session types
